+++ templates/book_seats.html
@@ -25,7 +25,7 @@
         <label for="ilosc-miejsc">Ilość miejsc:</label>
         <input type="number" id="ilosc-miejsc" name="ilosc-miejsc" min="1" max="10" onchange="updateSeatSelection()">
     </div>
-    <form id="seats-form" action="{{ url_for('summary') }}" method="post" onsubmit="return validateForm()">
+    <form id="seats-form" action="{{ url_for('user.summary') }}" method="post" onsubmit="return validateForm()">
         <input type="hidden" name="id_filmu" value="{{ id_filmu }}">
         <input type="hidden" name="id_kina" value="{{ id_kina }}">
         <input type="hidden" name="date" value="{{ date }}">
