+++ templates/payment.html
@@ -4,7 +4,7 @@
 <div class="container">
     <h1>Płatność</h1>
     <p>Czy zapłaciłeś?</p>
-    <form action="{{ url_for('payment_confirmation') }}" method="post">
+    <form action="{{ url_for('user.payment_confirmation') }}" method="post">
         <input type="hidden" name="reservation_id" value="{{ reservation_id }}">
         <input type="hidden" name="total_cost" value="{{ total_cost }}">
         <button type="submit" name="confirmation" value="yes" class="btn btn-success">Tak</button>
